move resource-api to new backend directory
