+++ backend/package.json
@@ -2,10 +2,10 @@
   "name": "scbirm-resource-api",
   "version": "0.1.0",
   "description": "This API returns information items in the SimCity:BuildIt game.",
-  "main": "src/server.js",
+  "main": "./server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon src/server.js"
+    "start": "nodemon ./server.js"
   },
   "author": "Nikita Latushka",
   "license": "ISC",
